requirements.txt
langgraph
langchain
langchain-community
langchain-groq
mcp
beautifulsoup4
requests
sentence-transformers
python-dotenv
tqdm
playwright
pdfplumber
streamlit